Plugins: categorize linux.modxview as a malware plugin
